This merges the consent and cookie preferences modal and has some improvements

+++ public/class-gdpr-public.php
@@ -44,6 +44,14 @@ class GDPR_Public {
 	private $version;
 
 	/**
+	 * Allowed HTML for wp_kses.
+	 * @since  1.1.0
+	 * @access private
+	 * @var    array   $allowed_html   The allowed HTML for wp_kses.
+	 */
+	private $allowed_html;
+
+	/**
 	 * Initialize the class and set its properties.
 	 *
 	 * @since  1.0.0
@@ -54,6 +62,13 @@ class GDPR_Public {
 	public function __construct( $plugin_name, $version ) {
 		$this->plugin_name = $plugin_name;
 		$this->version     = $version;
+		$this->allowed_html = array(
+			'a' => array(
+				'href' => true,
+				'title' => true,
+				'target' => true,
+			),
+		);
 	}
 
 	/**
@@ -80,66 +95,44 @@ class GDPR_Public {
 	}
 
 	/**
-	 * Prints the cookie bar for the end user to save the cookie settings.
+	 * Prints the privacy bar for the end user to save the consent and cookie settings.
 	 * @since  1.0.0
 	 * @author Fernando Claussen <fernandoclaussen@gmail.com>
 	 */
-	public function cookie_bar() {
-		if ( isset( $_COOKIE['gdpr_approved_cookies'] ) ) { // Input var okay.
+	public function privacy_bar() {
+		if ( isset( $_COOKIE['gdpr']['allowed_cookies'] ) ) { // Input var okay.
 			return;
 		}
 
-		$content = get_option( 'gdpr_cookie_banner_content', '' );
-		$tabs = get_option( 'gdpr_cookie_popup_content', array() );
-		$link_label = get_option( 'gdpr_cookie_banner_privacy_policy_link_label', '' );
+		$content             = get_option( 'gdpr_cookie_banner_content', '' );
+		$tabs                = get_option( 'gdpr_cookie_popup_content', array() );
+		$link_label          = get_option( 'gdpr_cookie_banner_privacy_policy_link_label', '' );
 		$privacy_policy_page = get_option( 'gdpr_privacy_policy_page', 0 );
 
 		if ( empty( $content ) || empty( $tabs ) ) {
 			return;
 		}
 
-		include plugin_dir_path( __FILE__ ) . 'partials/cookie-bar.php';
+		include plugin_dir_path( __FILE__ ) . 'partials/privacy-bar.php';
 	}
 
 	/**
-	 * The cookie preferences modal.
+	 * The privacy preferences modal.
 	 * @since  1.0.0
 	 * @author Fernando Claussen <fernandoclaussen@gmail.com>
 	 */
-	public function cookie_preferences() {
+	public function privacy_preferences_modal() {
 		$cookie_privacy_excerpt = get_option( 'gdpr_cookie_privacy_excerpt', '' );
+		$consent_types = get_option( 'gdpr_consent_types', array() );
 		$privacy_policy_page = get_option( 'gdpr_privacy_policy_page', 0 );
-		$approved_cookies = isset( $_COOKIE['gdpr_approved_cookies'] ) ? json_decode( wp_unslash( $_COOKIE['gdpr_approved_cookies'] ) ) : array();
+		$approved_cookies = isset( $_COOKIE['gdpr']['allowed_cookies'] ) ? json_decode( wp_unslash( $_COOKIE['gdpr']['allowed_cookies'] ) ) : array();
+		$user_consents = isset( $_COOKIE['gdpr']['consent_types'] ) ? json_decode( wp_unslash( $_COOKIE['gdpr']['consent_types'] ) ) : array();
 		$tabs = get_option( 'gdpr_cookie_popup_content', array() );
 		if ( empty( $tabs ) ) {
 			return;
 		}
 
-		include plugin_dir_path( __FILE__ ) . 'partials/cookie-preferences.php';
-	}
-
-	/**
-	 * The consents preferences modal.
-	 * @since  1.0.0
-	 * @author Fernando Claussen <fernandoclaussen@gmail.com>
-	 */
-	public function consents_preferences() {
-
-		if ( ! is_user_logged_in() ) {
-			return;
-		}
-
-		$consent_types = get_option( 'gdpr_consent_types', array() );
-		if ( empty( $consent_types ) ) {
-			return;
-		}
-
-		if ( is_user_logged_in() ) {
-			$user = wp_get_current_user();
-			$user_consents = get_user_meta( $user->ID, 'gdpr_consents' );
-		}
-
-		include plugin_dir_path( __FILE__ ) . 'partials/consent-preferences.php';
+		include plugin_dir_path( __FILE__ ) . 'partials/privacy-preferences-modal.php';
 	}
 
 	/**
@@ -161,32 +154,61 @@ class GDPR_Public {
 	}
 
 	/**
-	 * Update the user consents from the front end modal window.
-	 * @since  1.0.0
+	 * Update the user allowed cookies and types of consent.
+	 * If the user is logged in, we also save consent to user meta.
+	 * @since  1.1.0
 	 * @author Fernando Claussen <fernandoclaussen@gmail.com>
 	 */
-	public function update_consents() {
-		if ( ! isset( $_POST['nonce'] ) || ! wp_verify_nonce( sanitize_key( $_POST['nonce'] ), 'update_consents' ) ) {
-			wp_send_json_error( esc_html__( 'We could not verify the the security token. Please try again.', 'gdpr' ) );
+	public function update_privacy_preferences() {
+		if ( ! isset( $_POST['update-privacy-preferences-nonce'] ) || ! wp_verify_nonce( sanitize_key( $_POST['update-privacy-preferences-nonce'] ), 'update_privacy_preferences' ) ) {
+			wp_die( esc_html__( 'We could not verify the the security token. Please try again.', 'gdpr' ) );
 		}
 
-		if ( ! isset( $_POST['consents'] ) ) {
-			wp_send_json_error( esc_html__( "You can't disable all consents.", 'gdpr' ) );
+		if ( ! isset( $_POST['user_consents'] ) ) {
+			wp_die( esc_html__( "You need to at least consent to our Privacy Policy.", 'gdpr' ) );
 		}
 
-		$user = wp_get_current_user();
-		$user_consents = get_user_meta( $user->ID, 'gdpr_consents' );
+		$consents = array_map( 'sanitize_text_field', (array) $_POST['user_consents'] );
+		$cookies = isset( $_POST['approved_cookies'] ) ? array_map( 'sanitize_text_field', (array) $_POST['approved_cookies'] ) : array();
 
-		delete_user_meta( $user->ID, 'gdpr_consents' );
-
-		GDPR_Audit_Log::log( $user->ID, esc_html__( 'User updated consents from modal. These are the user consents after the save:', 'gdpr' ) );
-		foreach ( (array) $_POST['consents'] as $consent ) {
-			$consent = sanitize_text_field( wp_unslash( $consent ) );
-			add_user_meta( $user->ID, 'gdpr_consents', $consent );
-			GDPR_Audit_Log::log( $user->ID, $consent['name'] );
+		if ( ! empty( $cookies ) ) {
+			$approved_cookies = array();
+			foreach ( $cookies as $cookieArr ) {
+				$cookieArr = json_decode( wp_unslash( $cookieArr ) );
+				foreach ( $cookieArr as $cookie ) {
+					$approved_cookies[] = $cookie;
+				}
+			}
 		}
 
-		wp_send_json_success( $user_consents );
+		$cookies_as_json = json_encode( $approved_cookies );
+		$consents_as_json = json_encode( $consents );
+
+		setcookie( "gdpr[allowed_cookies]", $cookies_as_json, time() + YEAR_IN_SECONDS, "/" );
+		setcookie( "gdpr[consent_types]", $consents_as_json, time() + YEAR_IN_SECONDS, "/" );
+
+		if ( is_user_logged_in() ) {
+			$user = wp_get_current_user();
+			GDPR_Audit_Log::log( $user->ID, esc_html__( 'User updated their privacy preferences. These are the new approved cookies and consent preferences:', 'gdpr' ) );
+			if ( ! empty( $consents ) ) {
+				delete_user_meta( $user->ID, 'gdpr_consents' );
+				foreach ( $consents as $consent ) {
+					$consent = sanitize_text_field( wp_unslash( $consent ) );
+					add_user_meta( $user->ID, 'gdpr_consents', $consent );
+					GDPR_Audit_Log::log( $user->ID, 'Consent: ' . $consent );
+				}
+			}
+
+			if ( ! empty( $approved_cookies ) ) {
+				foreach ( $approved_cookies as $cookie ) {
+					GDPR_Audit_Log::log( $user->ID, 'Cookie: ' . $cookie );
+				}
+			}
+
+		}
+
+		wp_safe_redirect( esc_url_raw( wp_get_referer() ) );
+		exit;
 	}
 
 	/**
@@ -210,25 +232,11 @@ class GDPR_Public {
 			return;
 		}
 
-		$user = wp_get_current_user();
+		$user          = wp_get_current_user();
 		$user_consents = get_user_meta( $user->ID, 'gdpr_consents' );
 
 		if ( ! in_array( 'privacy-policy', (array) $user_consents ) ) {
-		?>
-			<div class="gdpr-consent-modal">
-				<div class="gdpr-consent-modal-content">
-					<h3><?php esc_html_e( 'Our Privacy Policy has been updated.', 'gdpr' ); ?></h3>
-					<h4><?php esc_html_e( 'To continue using the site you need to read the revised version and agree to the terms.', 'gdpr' ); ?></h4>
-					<div class="privacy-viewer">
-						<?php echo apply_filters( 'the_content', $page_obj->post_content ); ?>
-					</div>
-					<div class="gdpr-consent-buttons">
-						<a href="#" class="gdpr-agree" data-nonce="<?php echo wp_create_nonce( 'user_agree_with_terms' ) ?>"><?php esc_html_e( 'Agree', 'gdpr' ); ?></a>
-						<a href="#" class="gdpr-disagree" data-nonce="<?php echo wp_create_nonce( 'user_disagree_with_terms' ) ?>"><?php esc_html_e( 'Disagree', 'gdpr' ); ?></a>
-					</div>
-				</div>
-			</div>
-		<?php
+			include plugin_dir_path( __FILE__ ) . 'partials/reconsent-modal.php';
 		}
 	}
